Split DNA parser runtime
