+++ js/dna.js
@@ -1,13 +1,18 @@
 // @ts-check
-// dna.js — DNA raw data import: parser, storage, context assembly
-// Parses raw DNA files from Ancestry, 23andMe, MyHeritage, FTDNA, Living DNA
-// Runs the heavy parsing in a Web Worker (inline blob) to keep UI responsive
+// dna.js — DNA storage, context assembly, and UI orchestration
 import { state } from './state.js';
 import { escapeAttr, escapeHTML, hashString, showNotification } from './utils.js';
 import { saveImportedData } from './data.js';
 import { closeModalOverlay, openModalOverlay } from './modal-lifecycle.js';
 import { dnaActionAttrs, initDnaActionDelegates } from './dna-actions.js';
-import { findGenotypeInfo as findGenotypeInfoImpl, findSnpHint as findSnpHintImpl, sortAlleles } from './dna-genotype.js';
+import { findGenotypeInfo as findGenotypeInfoImpl, findGenotypeMatch, findSnpHint as findSnpHintImpl, sortAlleles } from './dna-genotype.js';
+import {
+  detectDNAFile,
+  isDNAFile,
+  isDNAFileByContent,
+  parseClinicalSnpReportTextWithTable,
+  parseDNAFileWithTable,
+} from './dna-parser.js';
 import {
   cacheDnaSnpTable, callDnaFileHandler, clearPendingDnaImport,
   confirmDnaDeleteDialog, getDnaRuntimeState, getPendingDnaImport,
@@ -29,175 +34,7 @@ import {
   resolveHaplogroup,
   setManualHaplogroup,
 } from './dna-mtdna.js';
-// ═══════════════════════════════════════════════
-// FORMAT DETECTION
-// ═══════════════════════════════════════════════
-// Detect DNA file by checking first few lines of text content
-export function detectDNAFile(text) {
-  const first = text.slice(0, 1500);
-  if (first.startsWith('#AncestryDNA')) return 'ancestry';
-  // MyHeritage's Low-pass WGS export (2025+ format) prepends a ##fileformat=MyHeritage
-  // comment block before the column header. Older array-based exports go straight
-  // to the RSID header — the comment-skipping CSV branch below catches both.
-  if (/^##fileformat=MyHeritage/im.test(first)) return 'csv';
-  if (first.includes('generated by 23andMe')) {
-    // 23andMe Mito/Y files have the same header (~900 bytes) — need 1500 bytes to see data lines
-    // Filter out truncated trailing line from the byte slice
-    const dataLines = first.split(/\r?\n/).filter(l => l.trim() && !l.startsWith('#') && (l.match(/\t/g) || []).length >= 3);
-    if (dataLines.length > 0 && dataLines.every(l => /^\S+\tMT\t/.test(l))) return '23andme-mito';
-    if (dataLines.length > 0 && dataLines.every(l => /^\S+\tY\t/.test(l))) return '23andme-y';
-    return '23andme';
-  }
-  if (first.includes('# Living DNA customer genotype data download file version')) return 'livingdna';
-  // Illumina GenomeStudio export (DNAEra and other clinical labs running Illumina arrays):
-  // BOM-prefixed [Header] block with GSGT Version, then [Data] block with comma-separated
-  // "Sample Name,SNP Name,Chr,Position,Allele1 - Plus,Allele2 - Plus" rows.
-  if (first.includes('GSGT Version,')) return 'illumina-gsgt';
-  // CSV formats — skip past leading "#"/"##" comment lines so MyHeritage's new
-  // multi-line header doesn't push the column-header line out of position 0.
-  const firstNonComment = first.split(/\r?\n/).find(l => l.trim() && !l.startsWith('#'))?.replace(/^\uFEFF/, '').trim() || '';
-  const compactHeader = firstNonComment.replace(/\s+/g, '');
-  if (/^"?rsid"?,"?gene"?,"?category"?,"?genotype"?,"?zygosity"?,"?risk_?allele"?,"?result"?/i.test(compactHeader)) return 'annotated-snp-report';
-  if (/^RSID,CHROMOSOME,POSITION,RESULT$/i.test(firstNonComment)) return 'csv'; // MyHeritage or FTDNA
-  if (/^"?RSID"?,/i.test(firstNonComment)) return 'csv'; // quoted variant
-  // Tab-separated: rsid\tchromosome\tposition\tgenotype (23andMe without header, or stripped)
-  if (/^rsid\tchromosome\tposition\tgenotype/i.test(firstNonComment)) return '23andme';
-  // Headerless: lines starting with rs followed by tab-separated fields (rsid\tchr\tpos\tAA)
-  const dataLines = first.split(/\r?\n/).filter(l => l.trim() && !l.startsWith('#'));
-  if (dataLines.length > 0 && /^rs\d+\t\d+\t\d+\t[ACGT\-]{1,2}$/i.test(dataLines[0].trim())) return '23andme';
-  // mtDNA CSV: short files with position+allele lines like "263G", "10398G"
-  if (dataLines.length > 0 && dataLines.length < 200 && dataLines.filter(l => /^\d+[ACGT]$/i.test(l.trim())).length >= dataLines.length * 0.8) return 'mtdna';
-  return null;
-}
-// Check if a dropped/selected file looks like a DNA raw data file (by name only — sync check)
-export function isDNAFile(file) {
-  if (!file) return false;
-  const name = file.name.toLowerCase();
-  // Common DNA file patterns
-  if (name.includes('ancestrydna')) return true;
-  if (name.includes('23andme')) return true;
-  if (name.includes('myheritage') && name.includes('dna')) return true;
-  if (name.includes('family_finder')) return true;
-  if (name.includes('livingdna') || name.includes('living_dna')) return true;
-  if (name.includes('genome') || name.includes('genotype') || name.includes('raw_dna') || name.includes('rawdna')) return true;
-  if (name.includes('mtdna') || name.includes('mt-dna') || name.includes('mt_dna')) return true;
-  if (name.includes('dnaera')) return true;
-  return false;
-}
-// Content-based DNA detection — reads first 1500 bytes (async, needs to see past 23andMe header)
-export async function isDNAFileByContent(file) {
-  try {
-    const header = await file.slice(0, 1500).text();
-    return detectDNAFile(header) !== null;
-  } catch { return false; }
-}
-// ═══════════════════════════════════════════════
-// WEB WORKER PARSER (inline blob)
-// ═══════════════════════════════════════════════
-const WORKER_CODE = `
-// DNA parser worker — receives { file, snpIds, format }
-// Posts back { matches, source, totalLines, format }
-
-self.onmessage = async function(e) {
-  const { file, snpIds, format } = e.data;
-  const text = await file.text();
-  const lines = text.split(/\\r?\\n/);
-  const lookupSet = new Set(snpIds);
-  const matches = {};
-  let totalData = 0;
-  let detectedFormat = format;
-  // Illumina GSGT files have a [Header] block that must be skipped — flip
-  // this true once we hit [Data] and discard the column-header line that
-  // follows it.
-  let illuminaInData = false;
-
-  for (let i = 0; i < lines.length; i++) {
-    const line = lines[i];
-    if (!line || line.startsWith('#')) continue;
-
-    // Skip header lines
-    const upper = line.toUpperCase();
-    if (upper.startsWith('RSID') || upper.startsWith('"RSID')) continue;
-
-    let rsid, genotype;
-
-    if (detectedFormat === 'illumina-gsgt') {
-      // Walk the [Header] block until [Data], then skip the column-header
-      // line ("Sample Name,SNP Name,Chr,Position,Allele1 - Plus,Allele2 - Plus")
-      // before any rows are eligible.
-      const trimmed = line.trim();
-      if (!illuminaInData) {
-        // Note the double-escape: the worker code is in a template literal,
-        // so doubled backslashes here render as single backslashes in the
-        // worker source, giving a regex that matches literal "[Data]".
-        if (/^\\[Data\\]/i.test(trimmed)) illuminaInData = true;
-        continue;
-      }
-      if (/^Sample Name/i.test(trimmed)) continue;
-      const parts = trimmed.split(',');
-      if (parts.length < 6) continue;
-      rsid = parts[1].trim();           // SNP Name
-      // Illumina probes wrap rsids many ways: seq-rs1234, GSA-rs1234,
-      // ilmnseq_rs1234_ilmnTOP_5AT, BOT-rs1234, dup-seq-rs1234, etc.
-      // Extract the bare rsid so a wrapped probe still hits the lookup.
-      const m = rsid.match(/(rs\\d+)/);
-      if (m) rsid = m[1];
-      const a1 = parts[4].trim();        // Allele1 - Plus
-      const a2 = parts[5].trim();        // Allele2 - Plus
-      if (a1 === '-' || a2 === '-' || a1 === '0' || a2 === '0') { totalData++; continue; }
-      genotype = a1 + a2;
-    } else if (detectedFormat === 'ancestry') {
-      // Tab-separated: rsid, chromosome, position, allele1, allele2
-      const parts = line.split('\\t');
-      if (parts.length < 5) continue;
-      rsid = parts[0].trim();
-      const a1 = parts[3].trim();
-      const a2 = parts[4].trim();
-      if (a1 === '0' || a2 === '0' || a1 === '-' || a2 === '-') { totalData++; continue; }
-      genotype = a1 + a2;
-    } else if (detectedFormat === '23andme' || detectedFormat === 'livingdna') {
-      // Tab-separated: rsid, chromosome, position, genotype
-      const parts = line.split('\\t');
-      if (parts.length < 4) continue;
-      rsid = parts[0].trim();
-      genotype = parts[3].trim();
-      if (genotype === '--' || genotype === '00') { totalData++; continue; }
-    } else if (detectedFormat === 'csv') {
-      // Comma-separated: RSID, CHROMOSOME, POSITION, RESULT (possibly quoted)
-      const parts = line.split(',').map(s => s.replace(/^"|"$/g, '').trim());
-      if (parts.length < 4) continue;
-      rsid = parts[0];
-      genotype = parts[3];
-      if (genotype === '--' || genotype === '00' || !genotype) { totalData++; continue; }
-    } else {
-      continue;
-    }
-
-    // Skip internal IDs (23andMe uses i-prefixed IDs for some markers)
-    if (!rsid.startsWith('rs')) { totalData++; continue; }
-
-    totalData++;
-
-    if (lookupSet.has(rsid) && !matches[rsid]) {
-      // First non-missing call wins. Illumina chips often have multiple
-      // probes for the same SNP (seq-rsX + seq-rsX.1 + seq-rsX.2). They
-      // usually agree, but if a later probe failed, keeping the first
-      // valid call is safer than overwriting it.
-      matches[rsid] = genotype;
-    }
-  }
-
-  self.postMessage({ matches, source: detectedFormat, totalLines: totalData, format: detectedFormat });
-};
-`;
-let _workerBlobUrl = null;
-function createWorker() {
-  if (!_workerBlobUrl) {
-    const blob = new Blob([WORKER_CODE], { type: 'application/javascript' });
-    _workerBlobUrl = URL.createObjectURL(blob);
-  }
-  return new Worker(_workerBlobUrl);
-}
+export { detectDNAFile, isDNAFile, isDNAFileByContent };
 // ═══════════════════════════════════════════════
 // PARSE DNA FILE
 // ═══════════════════════════════════════════════
@@ -245,225 +82,8 @@ function loadSNPTable({ forceFresh = false } = {}) {
   return _snpTablePromise;
 }
 
-function parseCsvLine(line) {
-  const values = [];
-  let current = '';
-  let inQuotes = false;
-  for (let i = 0; i < line.length; i++) {
-    const ch = line[i];
-    if (ch === '"') {
-      if (inQuotes && line[i + 1] === '"') {
-        current += '"';
-        i++;
-      } else {
-        inQuotes = !inQuotes;
-      }
-    } else if (ch === ',' && !inQuotes) {
-      values.push(current.trim());
-      current = '';
-    } else {
-      current += ch;
-    }
-  }
-  values.push(current.trim());
-  return values.map(v => v.replace(/^\uFEFF/, '').trim());
-}
-
-function normalizeReportHeader(value) {
-  return String(value || '')
-    .replace(/^\uFEFF/, '')
-    .trim()
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, '_')
-    .replace(/^_+|_+$/g, '');
-}
-
-function normalizeReportText(value) {
-  return String(value || '').trim().toLowerCase().replace(/\u03B5/g, 'e');
-}
-
-function normalizeReportAllele(value) {
-  const allele = String(value || '').trim().toUpperCase();
-  return /^[ACGT]$/.test(allele) ? allele : '';
-}
-
-function getReportCell(cells, index) {
-  return index >= 0 ? (cells[index] || '').trim() : '';
-}
-
-function isHeterozygousGenotype(genotype) {
-  return /^[ACGT]{2}$/.test(genotype) && genotype[0] !== genotype[1];
-}
-
-function isHomozygousGenotype(genotype) {
-  return /^[ACGT]{2}$/.test(genotype) && genotype[0] === genotype[1];
-}
-
-function pickAnnotatedCandidate(candidates, riskAllele, mode = 'any') {
-  if (riskAllele) {
-    const riskMatches = candidates.filter(g => mode === 'excludeRisk' ? !g.includes(riskAllele) : g.includes(riskAllele));
-    if (riskMatches.length === 1) return riskMatches[0];
-  }
-  return candidates.length === 1 ? candidates[0] : null;
-}
-
-function inferApoeReportGenotype(rsid, resultText) {
-  const match = normalizeReportText(resultText).replace(/\s+/g, '').match(/e([234])\/e([234])/);
-  if (!match) return null;
-  const pair = [match[1], match[2]].sort().join('/');
-  const lookup = {
-    '2/2': { rs429358: 'TT', rs7412: 'TT' },
-    '2/3': { rs429358: 'TT', rs7412: 'CT' },
-    '2/4': { rs429358: 'CT', rs7412: 'CT' },
-    '3/3': { rs429358: 'TT', rs7412: 'CC' },
-    '3/4': { rs429358: 'TC', rs7412: 'CC' },
-    '4/4': { rs429358: 'CC', rs7412: 'CC' },
-  };
-  return lookup[pair]?.[rsid] || null;
-}
-
-function inferAnnotatedReportGenotype(rsid, entry, cells, columns) {
-  const rawGenotype = getReportCell(cells, columns.genotype);
-  const rawZygosity = getReportCell(cells, columns.zygosity);
-  const rawRiskAllele = getReportCell(cells, columns.riskAllele);
-  const rawResult = getReportCell(cells, columns.result);
-  const direct = rawGenotype.trim().toUpperCase();
-  if (/^[ACGT]{1,2}$/.test(direct)) return direct;
-
-  const apoe = inferApoeReportGenotype(rsid, rawResult) || inferApoeReportGenotype(rsid, rawGenotype);
-  if (apoe) return apoe;
-
-  const keys = Object.keys(entry?.genotypes || {}).filter(g => /^[ACGT]{2}$/.test(g));
-  if (!keys.length) return null;
-
-  const riskAllele = normalizeReportAllele(rawRiskAllele);
-  const combined = `${normalizeReportText(rawGenotype)} ${normalizeReportText(rawZygosity)} ${normalizeReportText(rawResult)}`;
-  const hetero = keys.filter(isHeterozygousGenotype);
-  const homo = keys.filter(isHomozygousGenotype);
-
-  if (/ref\/ref|wildtype|homozygous reference|risk allele absent|protective allele absent|variant absent|non[-\s]?carrier|not (?:a )?carrier/.test(combined)) {
-    const reference = keys.filter(g => entry.genotypes[g]?.effect === 'none');
-    const referenceHomo = reference.filter(isHomozygousGenotype);
-    return pickAnnotatedCandidate(referenceHomo.length ? referenceHomo : reference, riskAllele, 'excludeRisk');
-  }
-
-  if (/\bhet\b|heterozygous|one risk copy|one copy|(?:^|[^a-z-])carrier\b(?!-)/.test(combined)) {
-    return pickAnnotatedCandidate(hetero, riskAllele);
-  }
-
-  if (/homozygous variant|two risk copies|risk homozygous/.test(combined)) {
-    const nonReferenceHomo = homo.filter(g => entry.genotypes[g]?.effect !== 'none');
-    return pickAnnotatedCandidate(nonReferenceHomo, riskAllele);
-  }
-
-  return null;
-}
-
-const DNA_COMPLEMENT = { A: 'T', T: 'A', C: 'G', G: 'C' };
-function reverseComplementGenotype(genotype) {
-  return String(genotype || '').toUpperCase().split('').reverse().map(ch => DNA_COMPLEMENT[ch] || ch).join('');
-}
-function findGenotypeMatch(entry, genotype) {
-  const raw = String(genotype || '').trim().toUpperCase();
-  if (!entry || !entry.genotypes || !/^[ACGT]{1,2}$/.test(raw)) return null;
-  const tries = [];
-  const push = (value) => { if (value && !tries.includes(value)) tries.push(value); };
-  push(raw);
-  if (raw.length === 2) push(raw[1] + raw[0]);
-  push(sortAlleles(raw));
-  const rc = reverseComplementGenotype(raw);
-  const alleles = new Set(Object.keys(entry.genotypes).join('').split(''));
-  const palindromic = alleles.size === 2 && ((alleles.has('A') && alleles.has('T')) || (alleles.has('C') && alleles.has('G')));
-  if (!palindromic) {
-    push(rc);
-    if (rc.length === 2) push(rc[1] + rc[0]);
-    push(sortAlleles(rc));
-  }
-  for (const key of tries) {
-    if (entry.genotypes[key] != null) return { key, info: entry.genotypes[key] };
-  }
-  return null;
-}
-
-function parseAnnotatedSnpReport(text, snpTable) {
-  const lines = text.split(/\r?\n/);
-  const headerIndex = lines.findIndex(line => line.trim() && !line.trim().startsWith('#'));
-  if (headerIndex < 0) return { matches: {}, source: 'annotated-snp-report', totalLines: 0, format: 'annotated-snp-report' };
-
-  const headers = parseCsvLine(lines[headerIndex]).map(normalizeReportHeader);
-  const columns = {
-    rsid: headers.indexOf('rsid'),
-    genotype: headers.indexOf('genotype'),
-    zygosity: headers.indexOf('zygosity'),
-    riskAllele: headers.indexOf('risk_allele'),
-    result: headers.indexOf('result'),
-  };
-  if (columns.rsid < 0 || columns.genotype < 0 || columns.result < 0) {
-    throw new Error('Unrecognized annotated SNP report format');
-  }
-
-  const matches = {};
-  let totalData = 0;
-  for (let i = headerIndex + 1; i < lines.length; i++) {
-    if (!lines[i].trim()) continue;
-    const cells = parseCsvLine(lines[i]);
-    const rsid = getReportCell(cells, columns.rsid).toLowerCase();
-    if (!/^rs\d+$/.test(rsid)) continue;
-    totalData++;
-    if (!snpTable?.[rsid] || matches[rsid]) continue;
-    const genotype = inferAnnotatedReportGenotype(rsid, snpTable[rsid], cells, columns);
-    if (genotype) matches[rsid] = genotype;
-  }
-  return { matches, source: 'annotated-snp-report', totalLines: totalData, format: 'annotated-snp-report' };
-}
-
-function parseClinicalSnpGenotypeTail(tail) {
-  const cleaned = String(tail || '').replace(/\s+/g, ' ');
-  const matches = [...cleaned.matchAll(/(?:^|[^A-Z])([ACGT]{2})(?=\s*(?:[–-]|—|\b|$))/g)];
-  if (!matches.length) return null;
-  // Lab PDFs often prefix result rows with timestamps like 11:48CC; the last
-  // allele pair near the zygosity prose is the actual genotype.
-  return matches[matches.length - 1][1];
-}
-
 export function parseClinicalSnpReportText(text, options = {}) {
-  const source = options.source || options.fileName || 'Clinical SNP report';
-  const snpTable = _snpTable;
-  const body = String(text || '').replace(/\u00A0/g, ' ');
-  const matches = {};
-  let totalLines = 0;
-  const rsMatches = [...body.matchAll(/\b(rs\d+)\b/gi)];
-  for (let i = 0; i < rsMatches.length; i++) {
-    const rsid = rsMatches[i][1].toLowerCase();
-    const entry = snpTable?.[rsid];
-    if (!entry || matches[rsid]) continue;
-    totalLines++;
-    const start = Math.max(0, rsMatches[i].index - 90);
-    const end = i + 1 < rsMatches.length ? Math.min(body.length, rsMatches[i + 1].index) : Math.min(body.length, rsMatches[i].index + 220);
-    const chunk = body.slice(start, end);
-    const genotype = parseClinicalSnpGenotypeTail(chunk);
-    if (!genotype) continue;
-    const match = findGenotypeMatch(entry, genotype);
-    if (!match) continue;
-    matches[rsid] = {
-      genotype,
-      normalizedGenotype: match.key,
-      gene: entry.gene,
-      variant: entry.variant,
-      category: entry.category,
-      markers: entry.markers || [],
-      effect: match.info.effect,
-      valence: match.info.valence,
-      note: match.info.note,
-      source: { type: options.type || 'report-text', label: source, fileName: options.fileName || null, rawText: chunk.trim().slice(0, 500) },
-    };
-  }
-  return {
-    matches,
-    source,
-    totalLines,
-    coverage: { found: Object.keys(matches).length, total: Object.keys(snpTable || {}).filter(k => k.startsWith('rs')).length },
-  };
+  return parseClinicalSnpReportTextWithTable(text, _snpTable, options);
 }
 
 // Eagerly load SNP table when genetics data exists (e.g. after JSON import)
@@ -489,57 +109,7 @@ export async function parseDNAFile(file) {
   // else (rendering, re-rendering, dashboard tooltips); the import path
   // is the one place a stale cache silently drops valid SNPs.
   const snpTable = await loadSNPTable({ forceFresh: true });
-  const snpIds = Object.keys(snpTable).filter(k => k.startsWith('rs'));
-
-  // Detect format from first chunk
-  const headerChunk = await file.slice(0, 1000).text();
-  const format = detectDNAFile(headerChunk);
-  if (!format) throw new Error('Unrecognized DNA file format');
-
-  let result;
-  if (format === 'annotated-snp-report') {
-    result = parseAnnotatedSnpReport(await file.text(), snpTable);
-  } else {
-    // Run parser in a fresh worker (prevents handler conflicts on concurrent parses)
-    const worker = createWorker();
-    result = await new Promise((resolve, reject) => {
-      const timeout = setTimeout(() => { worker.terminate(); reject(new Error('DNA parsing timed out')); }, 30000);
-      worker.onmessage = (e) => { clearTimeout(timeout); worker.terminate(); resolve(e.data); };
-      worker.onerror = (e) => { clearTimeout(timeout); worker.terminate(); reject(new Error(e.message)); };
-      worker.postMessage({ file, snpIds, format });
-    });
-  }
-
-  // Enrich matches with lookup table data. SNPs whose raw call doesn't resolve
-  // to any catalog genotype (after direct/reversed/sorted/RC fallbacks) are
-  // dropped silently — those are MyHeritage Low-pass WGS imputation noise
-  // (e.g. "CG" for a C/T locus) that we have no curated interpretation for.
-  // Surfacing them as "Genotype not in lookup" was misleading: the rsID *is*
-  // curated, only the specific allele call isn't a valid variant.
-  const enriched = {};
-  for (const [rsid, genotype] of Object.entries(result.matches)) {
-    const entry = snpTable?.[rsid];
-    if (!entry) continue;
-    const genotypeInfo = findGenotypeInfo(entry, genotype);
-    if (!genotypeInfo) continue;
-    enriched[rsid] = {
-      genotype,
-      gene: entry.gene,
-      variant: entry.variant,
-      category: entry.category,
-      markers: entry.markers || [],
-      effect: genotypeInfo.effect,
-      valence: genotypeInfo.valence,
-      note: genotypeInfo.note,
-    };
-  }
-
-  return {
-    matches: enriched,
-    source: formatSourceName(result.source),
-    totalLines: result.totalLines,
-    coverage: { found: Object.keys(enriched).length, total: snpIds.length },
-  };
+  return parseDNAFileWithTable(file, snpTable);
 }
 
 // Single source of truth for "given a raw genotype call, what catalog entry
@@ -554,11 +124,6 @@ export function findGenotypeInfo(entry, genotype) {
 // is correct here too.
 export function findSnpHint(entry, genotype) {
   return findSnpHintImpl(entry, genotype);
-}
-
-function formatSourceName(format) {
-  const names = { ancestry: 'AncestryDNA', '23andme': '23andMe', livingdna: 'Living DNA', csv: 'MyHeritage/FTDNA', 'illumina-gsgt': 'Illumina GenomeStudio (DNAEra)', 'annotated-snp-report': 'Annotated SNP report' };
-  return names[format] || format;
 }
 
 // ═══════════════════════════════════════════════
