+++ scripts/quality-baseline.json
@@ -2,5 +2,5 @@
   "inlineEventAttributes": 0,
   "windowReferences": 202,
   "largeJsFilesOver800Lines": 28,
-  "maxJsFileLines": 1511
+  "maxJsFileLines": 1496
 }
